refactor: share workspace helpers
